Move as_arithmetic
