+++ include/boost/simd/arch/common/generic/function/genmask.hpp
@@ -14,7 +14,7 @@
 
 #include <boost/simd/function/scalar/bitwise_cast.hpp>
 #include <boost/simd/sdk/hierarchy/logical.hpp>
-#include <boost/simd/as_arithmetic.hpp>
+#include <boost/simd/meta/as_arithmetic.hpp>
 #include <boost/simd/constant/allbits.hpp>
 #include <boost/simd/constant/zero.hpp>
 #include <boost/dispatch/function/overload.hpp>
