+++ include/boost/simd/arch/common/simd/function/make.hpp
@@ -16,7 +16,7 @@
 #include <boost/simd/function/load.hpp>
 #include <boost/simd/sdk/hierarchy/logical.hpp>
 #include <boost/simd/sdk/hierarchy/simd.hpp>
-#include <boost/simd/as_arithmetic.hpp>
+#include <boost/simd/meta/as_arithmetic.hpp>
 #include <boost/dispatch/function/overload.hpp>
 #include <boost/config.hpp>
 #include <initializer_list>
