+++ include/boost/simd/sdk/is_bitwise_logical.hpp
@@ -13,7 +13,7 @@
 #define BOOST_SIMD_SDK_IS_BITWISE_LOGICAL_HPP_INCLUDED
 
 #include <boost/simd/sdk/as_logical.hpp>
-#include <boost/simd/as_arithmetic.hpp>
+#include <boost/simd/meta/as_arithmetic.hpp>
 #include <type_traits>
 
 namespace boost { namespace simd
